gnu: qt: Update qt to 5.7.0.
* gnu/packages/qt.scm
(qtbase): Update it.
(qtsvg): Update it.
(qtimageformats): Update it.
(qtx11extras): Update it.
(qtxmlpatterns): Update it.
(qtdeclarative): Update it.
(qtconnectivity): Update it.
(qtwebsockets): Update it.
(qtsensors): Update it.
(qtmultimedia): Update it.
(qtwayland): Update it.
(qtserialport): Update it.
(qtwebchannel): Update it.
(qtlocation): Update it.
(qttools): Update it.
(qtscript): Update it.
(qtquickcontrols): Update it.
(qtquickcontrols2): Update it.
(qtgraphicaleffects): Update it.