gnu: Fix and improve the mongo-tools package.
The hash of the source for this package is wrong. Most probably because the
source was changed from a tarball to a git repository without updating the
hash. Fixing this seems to break the check phase, so I rewrote that as well,
making changes to the build and install phases as needed.

* gnu/packages/databases.scm (mongo-tools)[source]: Update the hash value.
  [arguments]: Remove the different unpack path, as this isn't necessary when
  using the git repository as a source. Move the list of tools to make it
  accessible from multiple phases. Rewrite the build phase, getting it to
  install the tools in to the location the tests seem to expect to find
  them. Replace the check phase to run the tests for each tool
  individually. Add an install phase to install the tools, now that go install
  is not being run in the build phase.
  [native-inputs]: Add go-github.com-smartystreets-goconvey as a native-input
  as this is required for running the tests.