pack: Use guile-for-build for the target system.
Until now, running "guix pack -s i686-linux" on an x86_64-linux machine,
for instance, would use an x86_64 guile for module derivations.  This
was OK until now, but would break when passing "--localstatedir" due to
the introduction of guile-sqlite3: we'd be using the i686 guile-sqlite3
along with the x86_64 guile.

* guix/scripts/pack.scm (guix-pack): Pass the 'system option from OPTS
to 'package-derivation'.