gnu: mes: Update to 0.18.0-08f04f55.
This update is a preparation for the %bootstrap-mes package; due to some small
problems it fails to build with plain mes-0.18.  mes-boot0 stays at 0.18 only
to avoid another rebuild world.

%bootstrap-mes brings the building and packaging the Mes bootstrap seeds from
manual operation into Guix.  We will bump mes and mes-boot0 to a future 0.18.1
or 0.19 and regenerate %bootstrap-mes in a nex iteration.

    08f04f55 build: Oops, remove stray lib/linux/x86_64-mes/crt1.
    33f37f27 build: x86_64 bootstrap build fixes and workaround.
    4ae2a111 doc: Release update.
    5277669b mescc: Oops, delete wrong line of assembly.
    44cc97a8 admin: Release update.

* gnu/packages/mes.scm (mes): Update to 0.18.0-08f04f55.
* gnu/packages/commencement.scm (mes-boot0): Stay at 0.18.0.