gnu: emacs-ggtags: Don't require GNU Global in path
emacs-ggtags relies on the GNU Global tools (global and gtags) to be available
in the path. This way it will grab some random executable if it will find
executables at all.
This is fixed by setting the prefix to GNU Global in the store. This way
emacs-ggtags will use specific deterministic executables.

* gnu/packages/emacs.scm (emacs-ggtags):
[inputs]: Add GNU Global to the inputs
[arguments]: Add a phase for substituting the right path to the executables

Signed-off-by: Ludovic Courtès <ludo@gnu.org>