gnu: linkchecker: Fix most test failures.
Add a patch to skip more tests that require the network. Along with some other
changes, this gets all but of one of the tests passing.

The remaining test is broken due to a problem with the Python package in Guix,
which is fixed by [1] which is currently on the core-updates branch. This
patch doesn't work on core-updates due to python2-miniboa not being present.

1: d453b0e11d3f001a8160a7d126fdbf40e45d5042

* gnu/packages/python.scm (linkchecker)[source]: Add patch.
[native-inputs]: Add more inputs required for the tests.
[arguments]: Enable the tests, move the 'check phase after 'install. Test
using the installed software, and use py.test.