gnu: python-2.7: Fix build.
The -j flag is passed in through #:make-flags and EXTRATESTOPTS. This is
combined with the TESTOPTS, and in Python 2, that includes the -l flag by
default.

The -l flag, is for finding memory leaks in tests, and is incompatible with
the -j flag, and the build previously failed at the start of the 'check
phase. Therefore, remove the -l flag from the Python 2 testsuite, as it
doesn't sound like a useful thing to do when building the package for Guix.

* gnu/packages/python.scm (python-2.7): Add 'remove-findleaks-from-testopts
phase.