gnu: gnome: Remove '--disable-acceleration-check' flag from 'gnome.desktop'.
With this change the GNOME session is now listed in the log-in screen of
GDM.

* gnu/packages/gnome.scm (gnome-session)[arguments]: Remove
'disable-hardware-acceleration-check' phase.  That phase was apparently
no longer needed.  Furthermore, GDM expects 'TryExec' to be a file name,
not a file name followed by arguments; thus, the option led GDM to
discard 'gnome.desktop' with "GdmSession: Command not found: TryExec".