gnu: MPD: Update to 0.21.4.
We no longer install any systemd unit files for MPD because this is no longer
supported upstream as a matter of policy or via MPD's build system. [0]

Specifically, they don't support using MPD with systemd unless MPD was linked
with libsystemd. The unit files are only built when this is the case.

[0]
<https://github.com/MusicPlayerDaemon/MPD/issues/124#issuecomment-334930456>

* gnu/packages/mpd.scm (mpd): Update to 0.21.4.
[build-system]: Use meson-build-system.
[native-inputs]: Add gcc-8, gcc-8:lib, and python-sphinx.
[arguments]: Add a 'expand-C++-include-path' phase. Remove the
'install-service-files' phase. Make sure to build the documentation.