gnu: ruby-safe-yaml: Add missing require 'time'.
Patch ruby-safe-yaml to fix an issue that would lead to an error like this:
uninitialized constant SafeYAML::Parse::Date::DateTime

It's been reported upstream [1], and this patch was taken from the upstream Git
repository.

1: https://github.com/dtao/safe_yaml/issues/80

* gnu/packages/ruby.scm (ruby-safe-yaml)[source]: Switch to the Git repository
so that applying patches works and add a patch.
[arguments]: Enable tests and add a phase to set the TZ environment variable,
as one of the tests depends on a certian timezone.
* gnu/packages/patches/ruby-safe-yaml-add-require-time.patch: New file.
* gnu/local.mk (dist_patch_DATA): Add new patch file.